requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-generativeai==0.5.4
google-cloud-logging==3.10.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
